error handling, HTTP response codes
adding ControllerAdvise impl
implementing http response codes

+++ src/main/java/com/example/demo/service/WorkoutService.java
@@ -1,5 +1,6 @@
 package com.example.demo.service;
 
+import com.example.demo.errorhandling.ResourceNotFoundException;
 import com.example.demo.model.Workout;
 import com.example.demo.repositories.WorkoutRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +24,11 @@ public class WorkoutService {
         return workoutRepository.findAllByUser();
     }
 
-    public void deleteWorkout(Integer id) {
-        workoutRepository.deleteById(id);
+    public boolean deleteWorkout(Integer id) {
+        if(workoutRepository.findById(id).isPresent()) {
+            workoutRepository.deleteById(id);
+            return true;
+        }
+        throw new ResourceNotFoundException(id);
     }
 }
